// src/types.ts
export type ConfluenceMacro = {
  name: string;
  id?: string | null;
  type: 'plain-text-body' | 'rich-text-body' | 'empty-body';
  params?: {
    filter: string[];
    defaults: Record<string, unknown>;
  };
  version?: number;
  token: Token;
};

export interface InlineAsyncData {
  type: string;
  pos: number;
  content: string;
  id: string;
}